galera: when containerized, use the host name rather than the bundle name
Galera resource agents records various attributes in the CIB that are
linked to the state of the local mysql database (e.g. last-committed).
When run in a bundle, the host name that the resource agent receives is
that of the bundle, but it cannot use it to stored attributes because
bundle replicas can move across physical hosts when restarted.

Use name of the node which is hosting the container to store to stores
attributes that reflect the state of the local mysql server. Keep
using the bundle's name for master/slave state attribute.