High: exportfs: Make fsid optional (bsc#1045161)
The fsid export option is not needed any more.  There
might be some strange corner cases where it can be
useful, but it definitely shouldn't be assumed to be
needed by default.

The `fsid=0` usage used to be important for NFSv4, but
this is no longer the case.  It is best *not* to use
`fsid=0` and to let `exportfs/mountd` create a
"NFSv4 pseudo root" themselves.  This is (almost)
always what is wanted and is least confusing.

The `fsid=N N!=0` usage was important some years ago
when the device number of hard drives started to change.
NFS used to use the device id, which was problematic.
Setting an explicit fsid= removed the instability.
However for quite some years now NFS has been using the
UUID of the filesystem, rather than the device id, to
identify a filesystem.  So unless you have multiple
filesystems with the same uuid, or are using some weird
type of filesystem which doesn't have a uuid, `fsid=`
isn't needed and is best avoided.

So exportfs_test should be changed to not require
`fsid=` any more.