LVM-activate: add new RA for LVM activation
You can see the motivation and basic design idea of
this RA from the header of LVM-activate. And, learn
the input parameters in "meta_data" func to use this
RA.

---
Changes since v1:
1. Comments typo fixes and enhancements.
2. Putting cur_systemid into extra_system_ids should work for
all sources of systemID.
3. Only stop lockspace of the shared VG when the final LV has
been deactivated.
4. Use 'dmsetup info -c -S "vgname=${VG} && lvname=${LV}"' to
check the LV status.

Changes since v2:
1. Check the status of all LVs in given VG.
2. Keep the existing tagging method in LVM-activate, but refer to it as
"tagging", dropping references to "host tags".
3. Use "/bin/sh".
4. Remove lvs in lvm_status(), though we have to make compromise that
the VG is assumably active if any LV of the VG is active.

Signed-off-by: Eric Ren <zren@suse.com>