LVM-activate: avoid unnecessary validation in stop action
"stop" action is very critical because stop failure can incur node
fencing, so we must try to stop sucessfully as possible as we can.

So, put lvm_validate() after lvm_status(), thus it won't validate
userlessly if the RA is already stopped. Also, removed the duplicate
checking for the VG.

Signed-off-by: Eric Ren <zren@suse.com>