Implement GTID/semi-sync based MariaDB Master/Slave resource-agent
The focus of this replication agent is to try to provide a setup
which will not lose data. This is done by tuning the semi-sync
configuration in such a way that only once data has been committed
on a slave will the master acknowledge a write.

During startup all nodes must be online and the node with the highest
GTID is used as master.

During failover it is assumed that only one node has failed and the
remaining node with the highest GTID is selected as the new master.