rabbitmq-cluster: get cluster status from mnesia during monitor
If mnesia is not running (for example if `rabbitmqctl stop_app` has
been called, or the service has paused during partition due to the
pause_minority strategy) then the cluster_status command to
rabbitmqctl will read the cached cluster status from disk and the
command returns 0 even though the service isn't really running at all.

Instead, force the cluster status to be read from mnesia.  If mnesia
is not running due to the above or similar circumstances, the command
will catch that and properly fail the monitor action.

Resolves: RHBZ#1595753