ldirectord: Remove /var/lock from systemd unit
The /var/lock/subsys/ directory is used in sysv init scripts, but are
not necessary in systemd services. If /var/lock/subsys/ does not exist,
the systemd service will fail to start due to the missing folder for the
ExecStartPost command. Since these locks are not required in systemd
services, we can safely remove them from the ldirectord unit file.