Low: NodeUtilization: Fix checking for Xen commands
Fix shell quoting in function Host_Total_Memory(). Missing quotes caused
that if no Xen command (xl or xm) was found the function made a
non-sense call to the info utility. This could have resulted in
unexpected messages appearing in pacemaker logs. Example:
> lrmd[3476]:   notice: prm_node_util_monitor_300000:3915:stderr [ info: Writing node (dir)Top... ]
> lrmd[3476]:   notice: prm_node_util_monitor_300000:3915:stderr [ info: Cannot find node `(dir)GNU Free Documentation License'. ]
> lrmd[3476]:   notice: prm_node_util_monitor_300000:3915:stderr [ info: Done. ]

The patch also changes a warning about no Xen command being available to
a debug message. With the fixed quoting, if a NodeUtilization resource
was configured to use the default utilization_hv_memory=true parameter
but the Xen tools were not available, the warning would be reported each
time when the monitor operation is run. This change preserves the
previous behaviour that reported no warning.