Filesystem: add trigger_udev_rules_if_need() for -U, -L, or /dev/xxx device
DEVICE parameter of this RA accepts "-U <uuid>" and "-L <label>" in
addition to "/dev/xxx". Let's add a new function
trigger_udev_rules_if_needed() to accommodate them all.

It is useful in the case a fresh filesystem is just created from another
node on the shared storage, and is not visible yet.

Signed-off-by: Roger Zhou <zzhou@suse.com>