Add new tickle_tcp implementation
This commit replaces the previous implementation of tickle_tcp
with a new implementation.
Apart from fixing various problems of the previous implementation,
the new implementation adds support for link-local IPv6 addresses
which was incomplete in the previous implementation (although
that does not seem really useful anyway), and it has a more
whitespace-tolerant parser.

This implementation should also be more portable than the
previous one, which relied on Linux-specific data structures.