Increase the security of monitor user in oracle
With static default credentials for a user is not safe
even for limited privilege. A local user may login to
the DB to do some attack by bugs or leaks.(#1049)

It is better to replace the automatic creation with
new user and leave the default password for legacy update.

Also an open issue (#1030) for this.