galera: Fix value used for connecting with the empty password.
When check_passwd_use_empty option was introduced, it added a possibility
to connect to mysql explicitly using empty password, thus bypassing defaults
set in ~/.my.cnf.
However the the empty password set for MYSQL_OPTIONS_CHECK variable was wrong,
adding too many levels of escaping and ultimately failing when passed to mysql client.

This commit changes the password value in MYSQL_OPTIONS_CHECK so that it really
ends in passing the empty string.