Low: galera: Fix handling of Galera nodes when they include a port number
When Galera is configured to use a non-default base port, this port
number needs to be specified also in wsrep_cluster_address. Example:
> wsrep_cluster_address="gcomm://bsc1170646-node1:6666,bsc1170646-node2:6666"

This causes a problem for the Galera resource agent because it splits
nodes in wsrep_cluster_address on character ',' and then expects to map
obtained Galera node names to Pacemaker node names. The mapping is done
either directly or indirectly via the cluster_host_map but fails in both
cases as cluster_host_map is also not expected to include any port
numbers. A result of the wrong mapping is that the agent invokes
'crm_attribute -N <node>' with a non-existent node name and is therefore
not able to obtain correct attribute values.

The patch addresses the problem by stripping any port number from a
Galera node in function galera_to_pcmk_name().