podman: make sure to remove containers with lingering exec sessions
It may happen that some "podman exec" commands don't finish
cleanly and leave lingering "Exec sessions" in the container's
state. In that case, a "podman rm" command will always fail.

To overcome the podman bug, issue a "podman rm -f" command when
we detect a container is stopped but still has some lingering
"Exec sessions" associated with it.

Related-Bug: rhbz#1839721