[gcp-vpc-move-vip] Improve accuracy of metadata and log messages
The metadata for the gcp-vpc-move-vip resource agent states that it
configures an alias IP address, when it actually configures an alias IP
address **range**.

The distinction between an alias IP address and an alias IP address
range is quite important. Since the RA configures an alias range,
**every** IP address within the subnet specified by the netmask is an
alias pointing to the node where the gcp-vpc-move-vip is running.

This patch corrects the metadata to make it clear that the RA manages
IP ranges. It also makes this correction for certain log messages.
Finally, it breaks lines to keep them under 80 characters.