ocf-distro: Try os-release first to get release ID
We continue to use lsb_release and the distro-specific /etc/*-release
files as a fallback if we don't find what we need in os-release.

os-release should be present on systemd-based systems and is intended to
be used as the standard, replacing lsb_release and the distro-specific
/etc/*-release files.[1]

It is necessary to implement os-release here in ocf-distro, for at least
two reasons:
  1. /etc/SuSE-release is deprecated and does not exist on modern
OpenSUSE. The existing get_release_id() may fail if lsb_release is not
installed.
  2. We currently have no way to determine if a system is Ubuntu unless
lsb_release is installed. This matters because the version info in
/etc/debian_version is incorrect for Ubuntu.

Related advantages of os-release:
  - The os-release file can give us distro information that we can't get
from merely checking the existence of files. Ubuntu and OpenSUSE are two
examples, as mentioned, but os-release should also work for more obscure
distros, provided that they use systemd.
  - There is no dependency on installing an optional package, as there
may be for lsb_release.

We need to make os-release the top-level, first-choice check because
it's generally the most specific check. For example, if we check
$_DEBIAN_VERSION_FILE before we check os-release, then we'll prematurely
declare that Ubuntu is Debian.

[1] 0pointer.de/blog/projects/os-release

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>