ocf-distro: Normalize OS release IDs in get_release_id()
This is to provide consistent output to the caller of get_release_id(),
regardless of the **source** of the information.

Strip off extraneous info (e.g., "Linux", "GNU/Linux", "release...",
whitespace) from OS release ID string. Then if it's a known OS release,
convert it to the same format that would be found in os-release.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>