ocf-distro: Use Solaris-compatible grep
Solaris's default grep doesn't accept an -E option for EREs or a -q
option for quiet. So we work around this by using -e options with BREs,
and redirection to /dev/null.

Also remove the distro regex constants from the top of the file, as
they're only used within the is_*_based() functions.

Reference:
https://docs.oracle.com/cd/E23823_01/html/816-5165/grep-1.html

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>