LVM-activate: Better handling of transiently missing PVs
This is a significant overhaul of the lvm_status() function to handle
situations where PVs are missing. A limitation of the current
lvm_status() is that it simply checks whether device mappings exist for
a VG. It doesn't run `vgs` to check whether the VG actually exists
(i.e., its PV is present). This can lead to false-success monitor
operations if the PV (and thus the VG) has gone missing. It can also
lead to stop failures if a PV has gone missing: lvm_status() returns
success because the device mappings are present, and then the RA fails
to deactivate the missing VG.

This also re-works the lvm_stop() function to specially handle a "PVs
are missing" error from lvm_status().

There are a couple of open questions in TODO's, but I hope this is a
more robust RA than what we had.

Resolves: RHBZ#1902433, RHBZ#1905825

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>