LVM-activate: Fix dd test to cope with mounted fs
If a filesystem is mounted but the underlying block device is
unavailable, then a dd test using the LV may still succeed. It can read
successfully from cache (seemingly up to 4 KiB before failing).

We want the dd test to fail if the block device is inaccessible, so here
we use iflag=direct.

However, iflag=direct fails with EINVAL if the block size is not a
multiple of 512, at least for ext4. See `info dd`. Here we change the bs
option from 1 to 4096 (matching the Filesystem RA) in case some other FS
type requires the dd block size to be in multiples of a larger number.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>