LVM-activate: enforce ACTIVATION_SKIP flag to avoid the race condition
This commit enforces ACTIVATION_SKIP flag to LV whenever RA LVM-activate
starts or stops.

The reason behind that is, for HA_LVM, auto-activation must be disabled
to avoid the race condition among multiple nodes. In the practice, often
this get forgotten and brings the unexpected trouble shooting effort. To
avoid that, the lv_attr 10th bit(ACTIVATION SKIP flag) is very handy to
simplify the user experience than to manage the out-of-band
auto_activation_volume_list option in lvm.conf. [1]

This involves two lvchange options:
-k,--setactivationskip y|n
-K,--ignoreactivationskip

ACTIVATION_SKIP flag of lv metadata is added since v2_02_99 and is
safe for the metadata compatibility.

[1] https://www.redhat.com/archives/linux-lvm/2020-November/msg00005.html

Signed-off-by: Roger Zhou <zzhou@suse.com>