NVMe-oF-Target: New RAs nvmet-subsystem nvmet-namespace nvmet-port
Three new resource agents that allow to operate NVMe-oF Targets.
You need to have a nvmet-subsystem RA first, followed by namespace(s)
and port(s). In order to enable failover you need to set the UUID
on the namespace. I guess it is a good idea to set the nguid (on
the namespace) and the serial (on the subsystem) as well.